+++ src/vs/base/common/verifier.ts
